Some work with dates
Fixed date comparisons from strings to date-objects
Equipment: Added a check that equipment intances date_added cannot be moved forward in time if it includes components. I could more specifically check the oldest added date of each component and condition the check with that, but this would require going through all rows and would be "expensive". This will suffice.